refactor: CaptionLabeledTextField 구조 개선 및 관련 UI 컴포넌트 리팩터링
- `CaptionLabeledField`를 `CaptionLabeledTextField`로 통합 및 파일명 변경
- `CaptionLabeledTextField`의 복잡한 파라미터를 제거하고 표준 입력 케이스에 집중하도록 단순화 (간격 8dp 고정, 색상 기본값 적용 등)
- `EditorMessageSection`에서 변경된 `CaptionLabeledTextField` 반영 및 멀티라인 입력(본문) 영역의 레이아웃 직접 구성
- `EditorSectionLabel` 상단 간격 및 일부 레이아웃 시맨틱스(`isTraversalGroup`) 조정
- `CaptionLabeledTextField` 프리뷰 추가 및 미사용 임포트 정리
